src/kcidb.py: filter nodes that don't have a status set to True
Don't send checkout nodes to KCIDB if the status is not True (or
"pass" with the new enum in the next schema version).  This is because
checkout nodes are first created with "pending" status until another
client checks out the kernel code and pushes a source tarball and sets
the status to "pass".

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>